6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,6 @@
-#include<stdio.h>
+#include <stdio.h>
 /**
- * main - Entry point
+ * main - A program that prints the size of various types on the computer
  *
  * Return: Always 0 (Success)
  */
@@ -11,10 +11,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-	printf("Size of a chair: %d byte(s)\n", sizeof(char));
-	printf("Size of an int: %d byte(s)\n", sizeof(int));
-	printf("Size of a long int: %d byte(s)\n", sizeof(long int));
-	printf("Size of a long long int: %d byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %i byte(s)\n", sizeof(float));
+	printf("Size of a chair: %d byte(s)\n",(unsigned long)sizeof(char));
+	printf("Size of an int: %d byte(s)\n",(unsigned long)sizeof(int));
+	printf("Size of a long int: %d byte(s)\n", (unsigned long)sizeof(long int));
+	printf("Size of a long long int: %d byte(s)\n", (unsigned long)sizeof(long long int));
+	printf("Size of a float: %i byte(s)\n", (unsigned long)sizeof(float));
 	return (0);
 }
